cron2mail--::: https://www.putorius.net/stop-cron-daemon-from-sending-email-for.html

+++ killevils.sh
@@ -1,12 +1,19 @@
 #!/bin/sh
 #logger KILLEVILS called...
-evill2kill=`cat /home/watchdoguser/evil-process-watchdog-killer/PIDTOKILL`
+WDHOME=/home/watchdoguser/evil-process-watchdog-killer
+PID2KILL=$WDHOME/PIDTOKILL
+if [ ! -f "$PID2KILL" ]; then
+    echo "$PID2KILL not exists - nothing to do"
+    exit 0
+fi
+
+evill2kill=`cat $PID2KILL`
 if [ $evill2kill > 30 ]; then
-        evilpid=`cat /home/watchdoguser/evil-process-watchdog-killer/firstPID.log`
+        evilpid=`cat $WDHOME/firstPID.log`
         logger "IWANTTOKILLITKILLIT!!!---$evill2kill ========="
         kill -9 $evill2kill
-        rm /home/watchdoguser/evil-process-watchdog-killer/PIDTOKILL
-	rm /home/watchdoguser/evil-process-watchdog-killer/top1.log
+        rm $WDHOME/PIDTOKILL
+	rm $WDHOME/top1.log
 fi
